[modularize] Move `_initialize_serialization` into `object_store_client`.
This is because we moved `serialization_context_map` out before and
`_initialize_serialization` depends on it.
